refactor: extract AI conversation logic into reusable hook
- Created useAIConversation hook to eliminate code duplication
- Extracted common streaming logic, PostHog tracking, and state management
- Updated AIAssistantModal and AIAssistantInterfaceSimple to use shared hook
- Reduced code duplication by ~200 lines while maintaining functionality
- Improved maintainability and consistency across AI components
